Convert `describeContent` to TypeScript.
