+++ packages/visual-editor/public/agent.kit.json
@@ -18390,7 +18390,7 @@
               "id": "runJavascript-0",
               "type": "runJavascript",
               "configuration": {
-                "code": "const run = function describeContent(inputs) {\n  const { template } = inputs;\n  const params = unique([...collectParams(textFromLLMContent(template))]);\n\n  const props = Object.fromEntries(\n    params.map((param) => [\n      toId(param),\n      {\n        title: toTitle(param),\n        description: `The value to substitute for the parameter \"${param}\"`,\n        type: \"string\",\n      },\n    ])\n  );\n\n  const $inputSchema = {\n    properties: {\n      context: {\n        type: \"array\",\n        title: \"Context in\",\n        examples: [],\n        items: {\n          type: \"object\",\n          behavior: [\"llm-content\"],\n        },\n        default: '[{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}]',\n        description: \"The optional incoming conversation context\",\n      },\n      template: {\n        type: \"object\",\n        title: \"Text\",\n        examples: [],\n        behavior: [\"llm-content\", \"config\"],\n        default: \"null\",\n        description:\n          \"(Optional) The text that will initialize or be added to existing conversation context. Use mustache-style {{params}} to add parameters.\",\n      },\n    },\n    type: \"object\",\n    required: [],\n  };\n\n  const $outputSchema = {\n    type: \"object\",\n    properties: {\n      context: {\n        type: \"array\",\n        title: \"Context out\",\n        examples: [],\n        items: {\n          type: \"object\",\n          behavior: [\"llm-content\"],\n        },\n        description:\n          \"The resulting context, created from the template and parameters.\",\n      },\n    },\n    required: [],\n  };\n\n  const required = params.map(toId);\n\n  return mergeSchemas($inputSchema, $outputSchema, props);\n\n  function mergeSchemas(inputSchema, outputSchema, properties) {\n    return {\n      inputSchema: {\n        ...inputSchema,\n        properties: {\n          ...inputSchema.properties,\n          ...properties,\n        },\n        required: [...(inputSchema.required || []), ...required],\n      },\n      outputSchema: outputSchema,\n    };\n  }\n\n  function toId(param) {\n    return `p-${param}`;\n  }\n\n  function toTitle(id) {\n    const spaced = id?.replace(/[_-]/g, \" \");\n    return (\n      (spaced?.at(0)?.toUpperCase() ?? \"\") +\n      (spaced?.slice(1)?.toLowerCase() ?? \"\")\n    );\n  }\n\n  function textFromLLMContent(content) {\n    return content?.parts.map((item) => item.text).join(\"\\n\") || \"\";\n  }\n\n  function unique(params) {\n    return Array.from(new Set(params));\n  }\n\n  function collectParams(text) {\n    if (!text) return [];\n    const matches = text.matchAll(/{{(?<name>[\\w-]+)}}/g);\n    return Array.from(matches).map((match) => match.groups?.name || \"\");\n  }\n};",
+                "code": "const run = function describeContent(inputs){const{template}=inputs;const params=unique([...collectParams(textFromLLMContent(template))]);const props=Object.fromEntries(params.map(param=>[toId(param),{title:toTitle(param),description:`The value to substitute for the parameter \"${param}\"`,type:\"string\"}]));const $inputSchema={properties:{context:{type:\"array\",title:\"Context in\",examples:[],items:{type:\"object\",behavior:[\"llm-content\"]},default:'[{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}]',description:\"The optional incoming conversation context\"},template:{type:\"object\",title:\"Text\",examples:[],behavior:[\"llm-content\",\"config\"],default:\"null\",description:\"(Optional) The text that will initialize or be added to existing conversation context. Use mustache-style {{params}} to add parameters.\"}},type:\"object\",required:[]};const $outputSchema={type:\"object\",properties:{context:{type:\"array\",title:\"Context out\",examples:[],items:{type:\"object\",behavior:[\"llm-content\"]},description:\"The resulting context, created from the template and parameters.\"}},required:[]};const required=params.map(toId);return mergeSchemas($inputSchema,$outputSchema,props);function mergeSchemas(inputSchema,outputSchema,properties){return{inputSchema:{...inputSchema,properties:{...inputSchema.properties,...properties},required:[...inputSchema.required||[],...required]},outputSchema}}__name(mergeSchemas,\"mergeSchemas\");function toId(param){return`p-${param}`}__name(toId,\"toId\");function toTitle(id){const spaced=id?.replace(/[_-]/g,\" \");return(spaced?.at(0)?.toUpperCase()??\"\")+(spaced?.slice(1)?.toLowerCase()??\"\")}__name(toTitle,\"toTitle\");function textFromLLMContent(content2){return content2?.parts.map(item=>{return\"text\"in item?item.text:\"\"}).join(\"\\n\")||\"\"}__name(textFromLLMContent,\"textFromLLMContent\");function unique(params2){return Array.from(new Set(params2))}__name(unique,\"unique\");function collectParams(text){if(!text)return[];const matches=text.matchAll(/{{(?<name>[\\w-]+)}}/g);return Array.from(matches).map(match=>match.groups?.name||\"\")}__name(collectParams,\"collectParams\")};",
                 "inputSchema": {
                   "type": "object",
                   "properties": {},
